Add files via upload
Move testFileOfSourceCode from the Data Files folder to your working directory to get the program to work, or rename the filename variable to point to its location. 

outputOfFirstPass file provided is a sample. If you edit testFileOfSourceCode, it will create a new version of outputOfFirstPass in your working directory. 

NOTE: THIS IS AN INITIAL VERSION FOR FEEDBACK. IT WILL REQUIRE MANY MORE REGULAR EXPRESSIONS TO WORK PROPERLY.